fix(opendc-sdk): renamed converter to factory & register multiple compute monitors
